+++ examples/hattip-app/vite.config.ts
@@ -1,5 +1,3 @@
-import { hattip } from "@hattip/vite";
-import { pages } from "vike-cloudflare";
 import vikeSolid from "vike-solid/vite";
 import vike from "vike/plugin";
 import { defineConfig } from "vite";
@@ -8,15 +6,5 @@ export default defineConfig({
   build: {
     target: "es2022",
   },
-  plugins: [
-    hattip(),
-    vike({}),
-    vikeSolid(),
-    pages({
-      server: {
-        kind: "hattip",
-        entry: "hattip-entry.ts",
-      },
-    }),
-  ],
+  plugins: [vike(), vikeSolid()],
 });