+++ examples/hono-app/vite.config.ts
@@ -1,6 +1,3 @@
-import devServer from "@hono/vite-dev-server";
-import adapter from "@hono/vite-dev-server/cloudflare";
-import { pages } from "vike-cloudflare";
 import vikeSolid from "vike-solid/vite";
 import vike from "vike/plugin";
 import { defineConfig } from "vite";
@@ -9,30 +6,5 @@ export default defineConfig({
   build: {
     target: "es2022",
   },
-  plugins: [
-    devServer({
-      entry: "hono-entry.ts",
-      adapter,
-
-      exclude: [
-        /^\/@.+$/,
-        /.*\.(ts|tsx|vue)($|\?)/,
-        /.*\.(s?css|less)($|\?)/,
-        /^\/favicon\.ico$/,
-        /.*\.(svg|png)($|\?)/,
-        /^\/(public|assets|static)\/.+/,
-        /^\/node_modules\/.*/,
-      ],
-
-      injectClientScript: false,
-    }),
-    vike({}),
-    vikeSolid(),
-    pages({
-      server: {
-        kind: "hono",
-        entry: "hono-entry.ts",
-      },
-    }),
-  ],
+  plugins: [vike(), vikeSolid()],
 });