+++ packages/vike-cloudflare/src/hono/index.ts
@@ -1,0 +1,3 @@
+// Re-exporting Hono so that it can be referenced in `hono-dev`
+export { Hono } from "hono";
+export * from "vike-server/hono";