+++ packages/vike-cloudflare/tsup.config.ts
@@ -5,7 +5,14 @@ export default defineConfig({
   target: "node18",
   format: "esm",
   dts: true,
-  entry: ["src/index.ts"],
+  entry: {
+    index: "src/plugins/index.ts",
+    config: "src/config.ts",
+    hono: "src/hono/index.ts",
+    "hono/serve": "src/hono/serve.ts",
+    hattip: "src/hattip/index.ts",
+    "hattip/serve": "src/hattip/serve.ts",
+  },
   clean: true,
   esbuildPlugins: [RawPlugin()],
 });