+++ CHANGELOG.md
@@ -1,3 +1,51 @@
+## [0.2.3](https://github.com/vikejs/vike-cloudflare/compare/v0.2.2...v0.2.3) (2025-04-05)
+
+
+### Bug Fixes
+
+* avoid type mismatch between different Vite versions ([#56](https://github.com/vikejs/vike-cloudflare/issues/56)) ([7471fd3](https://github.com/vikejs/vike-cloudflare/commit/7471fd342e90d57b2dd8c568d06e41ad2c80af5e))
+
+
+
+## [0.2.2](https://github.com/vikejs/vike-cloudflare/compare/v0.2.1...v0.2.2) (2025-03-27)
+
+
+### Bug Fixes
+
+* wrong resolve conditions applied in dev ([48d5fa7](https://github.com/vikejs/vike-cloudflare/commit/48d5fa70b2504eb9b4cba0f35e1a60f826f0c7e6))
+
+
+
+## [0.2.1](https://github.com/vikejs/vike-cloudflare/compare/v0.2.0...v0.2.1) (2025-03-26)
+
+
+### Bug Fixes
+
+* use configEnvironment instead of config. Fixes [#52](https://github.com/vikejs/vike-cloudflare/issues/52) ([#53](https://github.com/vikejs/vike-cloudflare/issues/53)) ([ac1272a](https://github.com/vikejs/vike-cloudflare/commit/ac1272accac90361be8d000865fcb22604420daf))
+
+
+
+# [0.2.0](https://github.com/vikejs/vike-cloudflare/compare/v0.1.7...v0.2.0) (2025-03-25)
+
+
+### Bug Fixes
+
+* make `vike-cloudflare` a real Vike extension ([#37](https://github.com/vikejs/vike-cloudflare/issues/37)) ([#39](https://github.com/vikejs/vike-cloudflare/issues/39)) ([a003432](https://github.com/vikejs/vike-cloudflare/commit/a0034326f9dac1e7de100b4f44d401cb76a77795))
+* use `vike@>=0.4.227` ([#50](https://github.com/vikejs/vike-cloudflare/issues/50)) ([6064d3b](https://github.com/vikejs/vike-cloudflare/commit/6064d3b621b78c1e6179311d0c3272b53f7d9e3c))
+
+
+### Features
+
+* use `vike-server` ([#45](https://github.com/vikejs/vike-cloudflare/issues/45)) ([c140354](https://github.com/vikejs/vike-cloudflare/commit/c14035456ae9b5f71e60ed3bb431d6d2e3fe0cd4))
+
+
+### BREAKING CHANGES
+
+* `vike-cloudflare` is now a Vike extension. Check [migrating from `0.1.x` to `0.2.x`](https://vike.dev/migration/vike-cloudflare)
+* Use `vike@0.4.227` or above.
+
+
+
 ## [0.1.7](https://github.com/vikejs/vike-cloudflare/compare/v0.1.6...v0.1.7) (2025-01-24)